backend/requirements.txt

fastapi
uvicorn[standard]
asyncpg
sentence-transformers
unstructured[pdf]
pydantic
python-dotenv
requests
fabric
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
